refactor: orginisation and cleanup of routes and pages (#140)
