+++ src/components/projects/A32NX/ExtendedFeatures.tsx
