+++ src/pages/legal/Legal.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+import { TermsOfService } from '../../components/legal/TermsOfService';
+import { PrivacyPolicy } from '../../components/legal/PrivacyPolicy';
+
+export const ToS: React.FC = () => (
+    <TermsOfService />
+);
+
+export const Privacy: React.FC = () => (
+    <PrivacyPolicy />
+);
